backend/pyproject.toml
[tool.poetry]
name = "twitter"
version = "0.1.0"
description = "twitter_clone"
authors = ["Diego Cavalcanti <cavalcantidiego@hotmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
docker = "^7.1.0"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
python-decouple = "^3.8"
factory-boy = "^3.3.1"
pytest = "^8.3.4"
pillow = "^11.0.0"
django-extensions = "^3.2.3"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
